Fix bug where a spread of 10.0 doesn't get properly stored
When determining which team was favored, the previous code checked to
see if the contained 0.0.  If it did, a negative would not be added to
the number.  The issue was that 10.0 also satisfied this condition so no
negative was being added.  New code now trims string before so an exact
match for 0.0 can be found and a negative will be correctly added.  Also
allows for excel sheet to format text as a number istead of strings so
formulas can be performed on the data.